backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
python-jose[cryptography]
pydantic
python-dotenv
passlib[bcrypt]
python-multipart